Fix filter results on saved recipes view
